Move General Info api routes to source directory
